# contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Chicago Crime API
  version: 0.1.0
  description: |
    Public HTTP API exposing dbt marts for the Chicago Crime DBMS dashboard.

    Contract owners:
      - Producer: backend agent
      - Consumers: frontend, qa, docs, architect

    This file is the source of truth for HTTP shapes. The FastAPI app must be
    kept in sync; CI runs `make contracts-validate` to detect drift.
  contact:
    name: backend-agent
    url: ../agents/backend/AGENTS.md
  license:
    name: MIT
servers:
  - url: http://localhost:8000
    description: Local dev (docker compose)
  - url: https://api.chicago-pipeline.local
    description: Reserved (no public deployment yet)
...
paths:
...
  /api/timeseries:
    get:
      tags: [timeseries]
      summary: Crimes over time at daily/weekly/monthly granularity
      operationId: getTimeseries
      parameters:
        - in: query
          name: granularity
          schema: { type: string, enum: [daily, weekly, monthly], default: daily }
        - in: query
          name: from_date
          schema: { type: string, format: date }
        - in: query
          name: to_date
          schema: { type: string, format: date }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/TimeseriesPoint" }
  /api/timeseries/forecast:
    get:
      tags: [timeseries]
      summary: Forecast band for the next N days
      operationId: getForecast
      parameters:
        - in: query
          name: horizon
          schema: { type: integer, default: 14, minimum: 1, maximum: 60 }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ForecastBundle" }
  /api/timeseries/anomalies:
    get:
      tags: [timeseries]
      summary: Days with z-score above threshold
      operationId: getAnomalies
      parameters:
        - in: query
          name: z
          schema: { type: number, default: 3 }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/AnomalyPoint" }
  /api/heatmap:
    get:
      tags: [timeseries]
      summary: 7x24 heatmap (weekday x hour)
      operationId: getHeatmap
      parameters:
        - in: query
          name: from_date
          schema: { type: string, format: date }
        - in: query
          name: to_date
          schema: { type: string, format: date }
        - in: query
          name: types
          schema: { type: string }
          description: Comma-separated primary types
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Heatmap" }
...
  /api/crime-types/top:
    get:
      tags: [crime-types]
      summary: Top N crime types by count
      operationId: getCrimeTypesTop
      parameters:
        - in: query
          name: from_date
          schema: { type: string, format: date }
        - in: query
          name: to_date
          schema: { type: string, format: date }
        - in: query
          name: types
          schema: { type: string }
          description: Comma-separated primary types
        - in: query
          name: limit
          schema: { type: integer, default: 10, maximum: 50 }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/CrimeTypeCount" }
  /api/crime-types/trend:
    get:
      tags: [crime-types]
      summary: Monthly trend for a specific primary type
      operationId: getCrimeTypeTrend
      parameters:
        - in: query
          name: type
          required: true
          schema: { type: string }
        - in: query
          name: from_date
          schema: { type: string, format: date }
        - in: query
          name: to_date
          schema: { type: string, format: date }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/TimeseriesPoint" }
...
components:
  schemas:
...
    TimeseriesPoint:
      type: object
      required: [ts, count]
      properties:
        ts: { type: string, format: date-time }
        count: { type: integer }
        arrests: { type: integer }
    ForecastBundle:
      type: object
      required: [history, forecast]
      properties:
        history:
          type: array
          items: { $ref: "#/components/schemas/TimeseriesPoint" }
        forecast:
          type: array
          items: { $ref: "#/components/schemas/ForecastPoint" }
    ForecastPoint:
      type: object
      required: [ts, yhat, yhat_lower, yhat_upper]
      properties:
        ts: { type: string, format: date-time }
        yhat: { type: number }
        yhat_lower: { type: number }
        yhat_upper: { type: number }
    AnomalyPoint:
      type: object
      required: [ts, z, count]
      properties:
        ts: { type: string, format: date-time }
        z: { type: number }
        count: { type: integer }
    Heatmap:
      type: object
      required: [matrix]
      properties:
        matrix:
          type: array
          items:
            type: array
            items: { type: integer }
          description: 7 rows (Mon..Sun) x 24 cols (0..23)
...
    CrimeTypeCount:
      type: object
      required: [primary_type, count]
      properties:
        primary_type: { type: string }
        count: { type: integer }
...